add mocks tests

+++ src/Ativos.Domain/Repositories/Usuarios/IUsuariosReadOnlyRepository.cs
@@ -5,7 +5,7 @@ public interface IUsuariosReadOnlyRepository
     Task<List<Entities.Usuario>> GetAll(long? matricula = null, string? nome = null, 
         string? departamento = null, string? cargo = null, string? role = null);
     Task<bool> ExistActiveUserMatricula(int matricula);
-    Task<Entities.Usuario?> GetUserByMatricula(int matricula);
+    Task<Entities.Usuario?> GetUserByMatricula(long matricula);
     Task<Entities.Usuario?> GetById(long id);
 
 }
